nova: add nova specific dataspace client
The dataspace client on base-nova stores the virtual address. If in the task
an pci_device is used and the nova systemcall 'assign_pci' succeeds, then we
know that an IOMMU is available. If the phys_addr of the dataspace is
requested, the nova specific dataspace client will return the stored virtual
address.