nova: delegate exception portals on CPU of caller
Before this commit the exception portals for new threads has been mapped
during the reply from the pager of the new thread. The pager thread may
be located on another CPU as the caller thread - initiation the creation of the
new thread - is running on, so mapping will be denied by the kernel.

Instead map the exception portals on the CPU of the caller thread by calling
its pager and take the pager cap of the new thread as (translate) argument.
The pager code use the translated cap to lookup the pager_object of the new
thread. With that object in hand the pager can map the right exception portals
during the reply.

Issue #814