tool: create image with grub with network enabled
image/iso_grub_tftp creates an iso image which contains just a grub 0.97-os
version with network support enabled. The image can be used to dd it to a
usb stick and to let it boot on a target machine. If the target machine
supports one of the configured grub network cards, it will request a dhcp
address, set up a tftp server and load a grub configfile from the tftp server.

With this commit some older machines without network PXE support can still be
used to easily load/test Genode setups.

Commandline of grub-0.97 "configure" call to create stage2_eltorito:
./configure --enable-e1000 --enable-rtl8139 --enable-ns8390 --enable-diskless -enable-via-rhine
make