intel_fb: request copy of Intel opregion - WIP
acpica and the Intel display driver tries to use the Intel opregion
simultaneously on Genode, which is not supported nor wanted for IO_MEM
(as which the opregion is handled).

Since it was not possible to remove it from acpica easily, as it is used
from ACPI AML code in SSDT tables, the patch adds the support to read in
the copy of the Intel opregion reported by the acpi_drv component.
The copy of the region was sufficient to make the Intel display driver happy
to lookup the Intel VBT (video bios table) information to setup all connectors
on a Fujitsu U7411 docking station.