sculpt: set fb_drv default config - TMP (Fixed by #4659)
TODO - On Sculpt first the intial fb_drv will be used, and later on, as
soon as the disks+filesystem is up, the actual user provided fb_drv will
be applied. The later re-configuration causes on some connectors DMA faults,
and the Intel fb driver will complain regular and spam the log. Even so the
displays work, the spamming will cause to hang up Sculpt after some time.