window_layouter: avoid suppressing gen_requests
If the dimension for a maximized client change from X -> Y -> X,
don't suppress to send update events to the client, as long as the window
layouter don't have seen the adjustment of the client to Y actually.
Otherwise, the client may stay at dimension Y, because the window layer
did not send an update for the 2. X dimension change.