sel4: eagerly switch fpu with vcpu on x86
Strange effects with FPU instructions during memset trigger aborts not only
in the VMM vCPU EP, but also in the nic or usb driver. When switching the
FPU for vCPUs eagerly, those effects seem to be gone. Presumably (but
there is no evidence) the lazy FPU switching in the kernel for vCPUs are
potentially not correct.