intel/gpu: recreate aperature for platform client
after ACPI resume also for the case, that the Intel display client
may establish a new platform session to this driver before the device
MMIO was re-acquired by this driver. The symptom manifests in failures of
the Intel display driver, when it tries to use the managed dataspace for the
sub-aperature MMIO range.

This commit is a fixup for

intel/gpu: allocate aperture on demand (Issue #5497)

Issue #5707